renamed email.py to email_client.py to fix run in pycharm
