+++ backend/src/emissions/email_client.py
